♻️Refactor(*): image preVeiw 함수로 분리
- tiptap.tsx 컴포넌트의 함수 길이가 너무 길어서 preView hook을 만들어서 분리했습니다.
